src/requirements.txt
Django==1.8.3
Markdown==2.6.2
django-bootstrap3==6.1.0
django-bootstrap3-datetimepicker==2.2.3
django-datetime-widget==0.9.3
django-extensions==1.5.5
django-filter==0.10.0
django-suit==0.2.13
djangorestframework==3.1.3
pytz==2015.4
requests==2.7.0
six==1.9.0
wsgiref==0.1.2